llm-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
debugpy>=1.8.0  # VS Code remote debugging
llama-cpp-python>=0.3.5
numpy<2.0.0

# Database Connectors
psycopg2-binary>=2.9.9  # PostgreSQL connector

# Vector & Graph Databases
neo4j>=5.26.0  # Primary: GraphRAG with Neo4j (Cypher 25 & VECTOR type support)
qdrant-client>=1.7.0  # Optional: Fallback vector DB

sentence-transformers==2.3.1
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
requests>=2.31.0
rapidfuzz>=3.0.0  # For fuzzy matching in query refinement
# chromadb>=0.5.20  # Deprecated

# LangGraph dependencies
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5

# PDF Processing & Document Parsing
--extra-index-url https://download.pytorch.org/whl/cu121
transformers>=4.30.0
torch==2.4.1+cu121
torchvision==0.19.1+cu121
pillow>=9.0.0
pdf2image>=1.16.0
accelerate>=0.20.0
loguru>=0.7.0
pypdfium2>=4.30.0
boto3>=1.34.0
reportlab>=4.0.0
opencv-python-headless>=4.10.0.84
beautifulsoup4>=4.12.3
pdfminer.six>=20231228

# PaddleOCR for Korean OCR (Apache 2.0, commercial use allowed)
# Note: Install paddlepaddle-gpu separately based on CUDA version
# pip install paddlepaddle-gpu==3.0.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu118/
paddleocr>=2.9.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Development & Linting
ruff>=0.4.0
mypy>=1.10.0
pre-commit>=3.7.0
